[flang][OpenMP] NFC: Move trait-matching helpers to Semantics
Move the five trait-matching functions that back metadirective lowering
out of Lower/OpenMP/Utils.cpp and into Semantics/openmp-utils:

  mapTraitSet, mapTraitSelector, evaluateUserCondition, getTraitScore,
  processTraitProperties

These functions only operate on parsed OpenMP trait selectors and the
Fortran evaluate/folding machinery and carry no MLIR or lowering
dependency. Placing them in Semantics makes them reusable from the
semantic-recording phase of `declare variant` (and any other feature
that needs to build a VariantMatchInfo without yet emitting MLIR).

In Utils.cpp, the four pure mapping/folding functions are removed
entirely; their callers now call the Semantics versions directly.
The processTraitProperties is kept as a local wrapper to generate a TODO
error on Clause and extension properties.
